Assistant checkpoint: Added Learn dropdown menu with requested options
Assistant generated file changes:
- client/src/components/header.tsx: Convert Learn link to dropdown menu with additional options

---

User prompt:

We currently have a link titled 'learn' in the header that links to our blog.   We'd like to turn that link into a drop down menu that includes these links:

Ask
Blog
Research
Search

The blog link in the dropdown should navigate to our existing blog page
The other pages are not yet created but will be soon.